refactor userController to split into controller and service and write tests
